+++ database/factories/UserFactory.php
@@ -28,16 +28,4 @@ class UserFactory extends Factory
             'password' => bcrypt('password'),
         ];
     }
-
-    /**
-     * Indicate that the model's email address should be unverified.
-     *
-     * @return static
-     */
-    public function unverified()
-    {
-        return $this->state(fn (array $attributes) => [
-            'email_verified_at' => null,
-        ]);
-    }
 }